quantile balancing moves into the model folder, the debug config turns it on, and the docstrings say what the functions do
The rule is Kimi K3's, so it sits in the model folder next to the router it
balances rather than in components, where core's model-agnostic sign-step hook
lives. The debug config sets it as the spec's post_optimizer_build_fn instead
of a second flavor. The docstrings drop the measurement records and the bin
sweeps -- those belong in the PR text -- and the module no longer names a file
outside the repository.
